refactor: Product 책임 이전을 위한 Name, Price 추가
